Extract shared chi-square special functions (#132)
The gamma / chi-square tail routines (log_gamma_lanczos,
continued_fraction_q, chi_square_ccdf) were copy-pasted in three places —
JarqueBera, LjungBox, and DistributionSelector — and had already drifted
(M_PI vs std::numbers::pi). DistributionSelector additionally wrapped the
tail in a chi_square_cdf = 1 - ccdf helper and then computed a p-value as
1 - chi_square_cdf(...), a double-negation round-trip.

Introduce a single ag::stats::SpecialFunctions translation unit exposing
log_gamma_lanczos and chi_square_ccdf, and route all three call sites
through it. The LR p-value in compareDistributions now calls
chi_square_ccdf directly, removing the double negation.
